+++ src/main/java/kitchenpos/application/OrderService.java
@@ -9,10 +9,10 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 import kitchenpos.application.exception.CustomIllegalArgumentException;
-import kitchenpos.dao.JpaMenuRepository;
+import kitchenpos.menu.domain.JpaMenuRepository;
 import kitchenpos.dao.JpaOrderRepository;
 import kitchenpos.dao.JpaOrderTableRepository;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.ui.dto.request.OrderRequest;
