+++ src/main/java/kitchenpos/menu/domain/JpaMenuRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.menu.domain;
 
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaMenuRepository extends JpaRepository<Menu,Long> {
