+++ src/main/java/kitchenpos/menu/ui/request/MenuRequest.java
@@ -1,10 +1,10 @@
-package kitchenpos.ui.dto.request;
+package kitchenpos.menu.ui.request;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
 
 public class MenuRequest {
