+++ src/test/java/kitchenpos/application/HappyOrderServiceTest.java
@@ -3,11 +3,11 @@ package kitchenpos.application;
 import java.math.BigDecimal;
 import java.util.Arrays;
 import kitchenpos.application.happy.HappyServiceTest;
-import kitchenpos.domain.Menu;
+import kitchenpos.menu.domain.Menu;
 import kitchenpos.domain.Order;
 import kitchenpos.ui.dto.request.MenuGroupRequest;
-import kitchenpos.ui.dto.request.MenuProductRequest;
-import kitchenpos.ui.dto.request.MenuRequest;
+import kitchenpos.menu.ui.request.MenuProductRequest;
+import kitchenpos.menu.ui.request.MenuRequest;
 import kitchenpos.ui.dto.request.OrderLineItemRequest;
 import kitchenpos.ui.dto.request.OrderRequest;
 import kitchenpos.ui.dto.request.OrderTableRequest;
