+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -10,13 +10,13 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import kitchenpos.dao.JpaMenuGroupRepository;
-import kitchenpos.dao.JpaMenuRepository;
+import kitchenpos.menu.domain.JpaMenuRepository;
 import kitchenpos.dao.JpaOrderRepository;
 import kitchenpos.dao.JpaOrderTableRepository;
 import kitchenpos.product.domain.JpaProductRepository;
 import kitchenpos.dao.JpaTableGroupRepository;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
